refactor: create team component
